pyproject.toml
[project]
name = "delivery-service"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.1",
    "fastapi>=0.120.1",
    "pydantic[email]>=2.12.3",
    "pydantic-settings>=2.11.0",
    "uvicorn>=0.38.0",
    "pyjwt[crypto]>=2.10.1",
]
